% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{linreg.coord}
\alias{linreg.coord}
\title{Linear regression with Coordinate descent}
\usage{
linreg.coord(Y, X)
}
\arguments{
\item{Y}{vector of outcome variable}

\item{X}{matrix of covariates}
}
\value{
coefs vector of coefficients
}
\description{
This function implements linear regression using the coordinate descent
algorithm. This minimises the sum of squares with respect to each parameter
at a time holding others fixed.
}
\examples{
linreg.coord(women$height,women$weight)
}
