% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{linreg.qr}
\alias{linreg.qr}
\title{Linear regression with Lapack's QR solver \code{dqrls}}
\usage{
linreg.qr(Y, X)
}
\arguments{
\item{Y}{vector of outcome variable}

\item{X}{matrix of covariates}
}
\value{
coefs vector of coefficients
}
\description{
This function implements linear regression using QR decomposition.
Lapack fortran subroutine \code{dqrls} is called in C for this purpose.
}
\examples{
linreg.qr(women$height,women$weight)
}
