% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{linreg.sor}
\alias{linreg.sor}
\title{Linear regression with the SOR Solver}
\usage{
linreg.sor(Y, X)
}
\arguments{
\item{Y}{vector of outcome variable}

\item{X}{matrix of covariates}
}
\value{
coefs vector of coefficients
}
\description{
This function implements linear regression by solving the normal equations. It
constructs matrices X'X and X'Y and solves for beta in X'X*beta=X'Y
using the iterative SOR solver of system of linear equations.
}
\examples{
linreg.sor(women$height,women$weight)
}
